requirements.txt
Flask
psycopg2-binary
python-dotenv
gunicorn
pandas_market_calendars
nselib
requests
